requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
requests>=2.31
python-dotenv>=1.0
typer>=0.12
tenacity>=8.2
rich>=13.7
